refactor: split host editor sections
